Ajustado retorno do imposto para double e adicionado Dockerfile.

+++ backend/src/main/java/com/projearq/sistemavendas/aplicacao/servicos/calculoImposto/CalculoImpostoArgentina.java
@@ -14,7 +14,7 @@ public class CalculoImpostoArgentina implements ICalculoImpostoStrategy {
     private final double IMPOSTO_IVA_PRODUTO_30L = 0.15;
 
     @Override
-    public Integer calculaValorDoImposto(List<ProdutoDTO> itens, double subtotal) {
+    public Double calculaValorDoImposto(List<ProdutoDTO> itens, double subtotal) {
 
         double imposto = 0;
         
@@ -29,6 +29,6 @@ public class CalculoImpostoArgentina implements ICalculoImpostoStrategy {
             }
         }
         
-        return (int) imposto;
+        return imposto;
     }
 }
